fix(sync): break budget estimator import cycle
